renaming library module
