+++ app/src/main/kotlin/com/intsoftdev/nre/MainActivity.kt
@@ -1,17 +1,11 @@
 package com.intsoftdev.nre
 
 import android.os.Bundle
-import android.util.Log
-import com.google.android.material.snackbar.Snackbar
 import androidx.appcompat.app.AppCompatActivity
 import android.view.Menu
 import android.view.MenuItem
-import androidx.lifecycle.Observer
-import com.intsoftdev.railsdk.data.StationModel
-import com.intsoftdev.railsdk.presentation.StationsViewModel
 
 import kotlinx.android.synthetic.main.activity_main.*
-import org.koin.android.viewmodel.ext.android.viewModel
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
