+++ app/src/main/kotlin/com/intsoftdev/nre/NREApplication.kt
@@ -2,7 +2,7 @@ package com.intsoftdev.nre
 
 import android.app.Application
 import com.intsoftdev.nre.di.viewModelModule
-import com.intsoftdev.railsdk.NRESDKInitialiser
+import com.intsoftdev.railclient.NRESDKInitialiser
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
