+++ app/src/main/kotlin/com/intsoftdev/nre/StationAdapter.kt
@@ -4,7 +4,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.intsoftdev.railsdk.data.StationModel
+import com.intsoftdev.railclient.data.StationModel
 import kotlinx.android.synthetic.main.station_row_layout.view.*
 import kotlin.properties.Delegates
