+++ app/src/main/kotlin/com/intsoftdev/nre/di/Modules.kt
@@ -1,6 +1,6 @@
 package com.intsoftdev.nre.di
 
-import com.intsoftdev.railsdk.presentation.StationsViewModel
+import com.intsoftdev.railclient.presentation.StationsViewModel
 import org.koin.android.viewmodel.dsl.viewModel
 import org.koin.dsl.module
