+++ libraries/nre-client/.gitignore
