+++ libraries/nre-client/proguard-rules.pro
