+++ libraries/nre-client/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.intsoftdev.railsdk">
+    package="com.intsoftdev.railclient">
 
     /
 </manifest>
