+++ libraries/nre-client/src/test/kotlin/com/intsoftdev/railclient/RxImmediateTestScheduler.kt
@@ -1,4 +1,4 @@
-package com.intsoftdev.railsdk
+package com.intsoftdev.railclient
 
 import androidx.annotation.NonNull
 import io.reactivex.Scheduler
@@ -6,7 +6,6 @@ import io.reactivex.android.plugins.RxAndroidPlugins
 import io.reactivex.disposables.Disposable
 import io.reactivex.internal.schedulers.ExecutorScheduler
 import io.reactivex.plugins.RxJavaPlugins
-import io.reactivex.schedulers.Schedulers
 import org.junit.rules.TestRule
 import org.junit.runner.Description
 import org.junit.runners.model.Statement
